#85 pom -> change dependencies. RedisService -> add validation for value and list. ValidationUtil -> create validation for value and list. RedisServiceTest -> add tests for null values

+++ redis-layer/src/main/java/team/asd/client/RedisClientImpl.java
@@ -10,55 +10,55 @@ import java.util.Map;
 
 @Repository
 public class RedisClientImpl implements RedisClient {
-	private final Jedis jedis;
+    private final Jedis jedis;
 
-	public RedisClientImpl(@Autowired JedisPool jedisPool) {
-		this.jedis = jedisPool.getResource();
-	}
+    public RedisClientImpl(@Autowired JedisPool jedisPool) {
+        this.jedis = jedisPool.getResource();
+    }
 
-	@Override
-	public String readByKey(String key) {
-		return jedis.get(key);
-	}
+    @Override
+    public String readByKey(String key) {
+        return jedis.get(key);
+    }
 
-	@Override
-	public String saveValueByKey(String key, String value) {
-		return jedis.set(key, value);
-	}
+    @Override
+    public String saveValueByKey(String key, String value) {
+        return jedis.set(key, value);
+    }
 
-	@Override
-	public Long saveList(String keyList, List<String> list) {
-		list.forEach(value -> jedis.rpush(keyList, value));
-		return (long) list.size();
-	}
+    @Override
+    public Long saveList(String keyList, List<String> list) {
+        list.forEach(value -> jedis.rpush(keyList, value));
+        return (long) list.size();
+    }
 
-	@Override
-	public Long saveElementIntoList(String keyList, String value) {
-		return jedis.rpush(keyList, value);
-	}
+    @Override
+    public Long saveElementIntoList(String keyList, String value) {
+        return jedis.rpush(keyList, value);
+    }
 
-	@Override
-	public List<String> retrieveList(String keyList) {
-		return jedis.lrange(keyList, 0, -1);
-	}
+    @Override
+    public List<String> retrieveList(String keyList) {
+        return jedis.lrange(keyList, 0, -1);
+    }
 
-	@Override
-	public Long saveValueInHashMap(String primaryKey, String secondaryKey, String value) {
-		return jedis.hset(primaryKey, secondaryKey, value);
-	}
+    @Override
+    public Long saveValueInHashMap(String primaryKey, String secondaryKey, String value) {
+        return jedis.hset(primaryKey, secondaryKey, value);
+    }
 
-	@Override
-	public String retrieveValueFromHashMap(String primaryKey, String secondaryKey) {
-		return jedis.hget(primaryKey, secondaryKey);
-	}
+    @Override
+    public String retrieveValueFromHashMap(String primaryKey, String secondaryKey) {
+        return jedis.hget(primaryKey, secondaryKey);
+    }
 
-	@Override
-	public Map<String, String> retrieveValueFromHashMap(String primaryKey) {
-		return jedis.hgetAll(primaryKey);
-	}
+    @Override
+    public Map<String, String> retrieveValueFromHashMap(String primaryKey) {
+        return jedis.hgetAll(primaryKey);
+    }
 
-	@Override
-	public String saveValueWithExpireDate(String key, String value, long expireDate) {
-		return jedis.setex(key, expireDate, value);
-	}
+    @Override
+    public String saveValueWithExpireDate(String key, String value, long expireDate) {
+        return jedis.setex(key, (int) expireDate, value);
+    }
 }
